Fixed issue 429. Also occuring was a buffer underflow (player ID of -1 was being used until actual player was assigned).
git-svn-id: http://bwapi.googlecode.com/svn@4058 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ branches/bwapi4/bwapi/BWAPI/Source/BWAPI/GameInternals.cpp
@@ -358,7 +358,7 @@ namespace BWAPI
 	}
 	int GameImpl::_currentPlayerId()
 	{
-		return (int)*BW::BWDATA_g_LocalHumanID;
+		return *BW::BWDATA_g_LocalHumanID;
 	}
 	bool GameImpl::tournamentCheck(int type, void *parameter)
 	{
